Lazy load commit data
This version took a big performance hit whenver I added shortstat to it. To
counteract that I am only loading in the SHA first and fetching all extra data
after the commit has been selected. This simplifies the commit parsing, allows
for more flexibility with data retrieved and reduces memory footprint of the
game (especially relevant in large repos).